cp16. jooq

+++ src/main/java/com/example/bank/dao/CardRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CardRepository extends JpaRepository<Card, Long> {
+public interface CardRepository extends JpaRepository<Card, Integer> {
 
-    List<Card> findAllByUser_UserId(Long userId);
+    List<Card> findAllByUser_UserId(Integer userId);
 
 }
