+++ src/main/java/com/example/bank/dao/TransferRepository.java
@@ -9,7 +9,7 @@ import java.util.Date;
 import java.util.List;
 
 @Repository
-public interface TransferRepository extends JpaRepository<Transfer, Long> {
+public interface TransferRepository extends JpaRepository<Transfer, Integer> {
 
     Transfer findTransferByToken(String token);
