+++ src/main/java/com/example/bank/service/TransactionsService.java
@@ -8,11 +8,11 @@ import java.util.List;
 
 public interface TransactionsService {
 
-    List<TransactionDTO> getTransactionsOfUser(Long userIdFrom, Long amountFrom, Long amountTo,
-                                            Date dateFrom, Date dateTo);
+    List<TransactionDTO> getTransactionsOfUser(int userIdFrom, Long amountFrom, Long amountTo,
+                                               Date dateFrom, Date dateTo);
 
-    List<TransactionDTO> getTransactionsOfUserToUser(Long cardIdFrom, Long cardIdTo, Long amountFrom,
-                                                  Long amountTo, Date dateFrom, Date dateTo);
+    List<TransactionDTO> getTransactionsOfUserToUser(int cardIdFrom, Long cardIdTo, Long amountFrom,
+                                                     Long amountTo, Date dateFrom, Date dateTo);
 
     void saveTransaction(Card cardFrom, Card cardTo, Long amount, Date date);
 }
